# app/docs/swagger.yaml
basePath: /
definitions:
...
  CalorieGuide-db_internal_meal.Meal:
    description: Приём пищи
    properties:
      author_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      likes:
        type: integer
      meal_name:
        type: string
      picture:
        items:
          type: integer
        type: array
      products_id:
        items:
          $ref: '#/definitions/CalorieGuide-db_internal_meal.Product'
        type: array
      total_calories:
        type: integer
      total_carbohydrates:
        type: integer
      total_fats:
        type: integer
      total_proteins:
        type: integer
    type: object
  CalorieGuide-db_internal_meal.Product:
    description: Продукт в приёме пищи
    properties:
      product_id:
        type: integer
      quantity:
        type: integer
    type: object
  CalorieGuide-db_internal_meal.WithLike:
    description: Приём пищи с лайком
    properties:
      author_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      isLiked:
        type: boolean
      likes:
        type: integer
      meal_name:
        type: string
      picture:
        items:
          type: integer
        type: array
      products_id:
        items:
          $ref: '#/definitions/CalorieGuide-db_internal_meal.Product'
        type: array
      total_calories:
        type: integer
      total_carbohydrates:
        type: integer
      total_fats:
        type: integer
      total_proteins:
        type: integer
    type: object
...
  internal_meal_handlers.FindAllRequest:
    properties:
      sort:
        type: string
      two-decade:
        type: integer
      user_id:
        type: integer
    type: object
  internal_meal_handlers.FindAllResponse:
    properties:
      meals:
        items:
          $ref: '#/definitions/CalorieGuide-db_internal_meal.WithLike'
        type: array
    type: object
  internal_meal_handlers.LikeResponse:
    properties:
      action:
        type: string
      meal_id:
        type: integer
      user_id:
        type: integer
    type: object
  internal_meal_handlers.SearchRequest:
    properties:
      user:
        type: integer
      word:
        type: string
    type: object
host: localhost:8090
info:
  contact: {}
  description: API для работы с продуктами, пользователями и приёмами пищи
  title: CalorieGuide API
  version: "1.0"
paths:
...
  /meal:
    post:
      consumes:
      - application/json
      description: Создает новый приём пищи
      parameters:
      - description: Данные приёма пищи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
        "400":
          description: Ошибка при создании
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Добавить новый приём пищи
      tags:
      - meals
  /meals:
    post:
      consumes:
      - application/json
      description: Возвращает список приёмов пищи с возможностью сортировки
      parameters:
      - description: Параметры сортировки
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_meal_handlers.FindAllRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_meal_handlers.FindAllResponse'
        "400":
          description: Ошибка при получении
          schema:
            type: string
      summary: Получить список приёмов пищи
      tags:
      - meals
  /meals/{id}:
    delete:
      description: Удаляет приём пищи по ID
      parameters:
      - description: ID приёма пищи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
        "400":
          description: Ошибка при удалении
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Удалить приём пищи
      tags:
      - meals
    get:
      description: Возвращает приём пищи по его ID
      parameters:
      - description: ID приёма пищи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
        "400":
          description: Приём пищи не найден
          schema:
            type: string
      summary: Получить приём пищи по ID
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Обновляет приём пищи по ID
      parameters:
      - description: ID приёма пищи
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные приёма пищи
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CalorieGuide-db_internal_meal.Meal'
        "400":
          description: Ошибка при обновлении
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Обновить приём пищи
      tags:
      - meals
  /meals/like:
    post:
      consumes:
      - application/json
      description: Ставит или убирает лайк приёму пищи
      parameters:
      - description: Данные лайка
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_meal_handlers.LikeResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_meal_handlers.LikeResponse'
        "400":
          description: Ошибка при лайке
          schema:
            type: string
        "401":
          description: Нет доступа
          schema:
            type: string
      summary: Лайк приёма пищи
      tags:
      - meals
  /meals/search:
    post:
      consumes:
      - application/json
      description: Ищет приёмы пищи по ключевому слову
      parameters:
      - description: Поисковый запрос
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_meal_handlers.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_meal_handlers.FindAllResponse'
        "400":
          description: Ошибка при поиске
          schema:
            type: string
      summary: Поиск приёмов пищи
      tags:
      - meals
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"